Minor fixes.
Bad file key generation logic. A warning. A possible type cast error and
trying to fix block table corruption (not yet fully fixed).

+++ src/main/java/systems/crigges/jmpq3/security/CryptographicLUT.java
@@ -44,7 +44,7 @@ final class CryptographicLUT {
             for (int table = 0; table < TABLE_NUMBER; table++) {
                 final short seed1 = (short) (seed = updateSeed(seed));
                 final short seed2 = (short) (seed = updateSeed(seed));
-                CRYPTOGRAPHIC_TABLES[table][value] = seed1 << 16 | Short.toUnsignedInt(seed2);
+                CRYPTOGRAPHIC_TABLES[table][value] = (int) seed1 << 16 | Short.toUnsignedInt(seed2);
             }
         }
     }
